Work on async UDFs and UDF refactoring
