+++ crates/arroyo-df/Cargo.toml
@@ -12,6 +12,7 @@ arroyo-connectors = { path = "../arroyo-connectors" }
 arroyo-formats = { path = "../arroyo-formats" }
 arroyo-operator = { path = "../arroyo-operator" }
 arroyo-storage = { path = "../arroyo-storage" }
+arroyo-udf-host = { path = "../arroyo-udf/arroyo-udf-host" }
 
 datafusion = "36.0"
 datafusion-expr = "36.0"
