+++ crates/arroyo-operator/Cargo.toml
@@ -13,6 +13,7 @@ arroyo-state = { path = "../arroyo-state" }
 arroyo-types = { path = "../arroyo-types" }
 arroyo-datastream = { path = "../arroyo-datastream" }
 arroyo-storage = { path = "../arroyo-storage" }
+arroyo-udf-common = { path = "../arroyo-udf/arroyo-udf-common" }
 
 anyhow = "1.0.71"
 arrow = { workspace = true, features = ["ffi"] }
