+++ crates/arroyo-operator/src/udfs.rs
@@ -216,127 +216,7 @@ fn scalar_none(datatype: &DataType) -> ScalarValue {
     }
 }
 
-#[derive(WrapperApi)]
-pub struct UdfDylibInterface {
-    run: unsafe extern "C-unwind" fn(
-        args_ptr: *mut FfiArraySchemaPair,
-        args_len: usize,
-        args_capacity: usize,
-    ) -> FfiArrayResult,
-}
 
-#[derive(WrapperApi)]
-pub struct AsyncUdfDylibInterface {
-    run: unsafe extern "C-unwind" fn(
-        args_ptr: *mut FfiArraySchemaPair,
-        args_len: usize,
-        args_capacity: usize,
-    ) -> FfiFuture<FfiArraySchemaPair>,
-}
-
-#[derive(Clone)]
-pub enum UdfInterface {
-    Sync(Arc<Container<UdfDylibInterface>>),
-    Async(Arc<Container<AsyncUdfDylibInterface>>),
-}
-
-#[derive(Clone)]
-pub struct UdfDylib {
-    name: Arc<String>,
-    signature: Arc<Signature>,
-    return_type: Arc<DataType>,
-    udf: UdfInterface,
-}
-
-impl UdfDylib {
-    pub fn new(name: String, signature: Signature, return_type: DataType, udf: UdfInterface) -> Self {
-        Self {
-            name: Arc::new(name),
-            signature: Arc::new(signature),
-            return_type: Arc::new(return_type),
-            udf,
-        }
-    }
-}
-
-pub struct SyncUdfDylib {
-    name: Arc<String>,
-    signature: Arc<Signature>,
-    return_type: Arc<DataType>,
-    udf: Arc<Container<UdfDylibInterface>>,
-}
-
-impl Debug for SyncUdfDylib {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("UdfDylib").finish()
-    }
-}
-
-impl TryFrom<&UdfDylib> for SyncUdfDylib {
-    type Error = anyhow::Error;
-
-    fn try_from(value: &UdfDylib) -> std::result::Result<Self, Self::Error> {
-        let UdfInterface::Sync(udf) = &value.udf else {
-            bail!("UDF is async but expected sync")
-        };
-
-        Ok(Self {
-            name: value.name.clone(),
-            signature: value.signature.clone(),
-            return_type: value.return_type.clone(),
-            udf: udf.clone(),
-        })
-    }
-}
-
-pub struct AsyncUdfDylib {
-    name: Arc<String>,
-    signature: Arc<Signature>,
-    return_type: Arc<DataType>,
-    udf: Arc<Container<AsyncUdfDylibInterface>>,
-}
-
-impl TryFrom<&UdfDylib> for AsyncUdfDylib {
-    type Error = anyhow::Error;
-
-    fn try_from(value: &UdfDylib) -> std::result::Result<Self, Self::Error> {
-        let UdfInterface::Async(udf) = &value.udf else {
-            bail!("UDF is sync but expected async")
-        };
-
-        Ok(Self {
-            name: value.name.clone(),
-            signature: value.signature.clone(),
-            return_type: value.return_type.clone(),
-            udf: udf.clone(),
-        })
-    }
-}
-
-
-impl AsyncUdfDylib {
-    pub async fn invoke(&self, args: Vec<ScalarValue>) -> ScalarValue {
-        let args: Vec<_> = args
-            .into_iter()
-            // We convert the scalar value to an array so we can ue the pre-existing FFI, but it
-            // may be more efficient with a new scalar FFI
-            .map(|arg| to_ffi(&arg.to_array_of_size(1).unwrap().to_data()).unwrap())
-            .map(|(data, schema)| FfiArraySchemaPair(data, schema))
-            .collect();
-
-        let len = args.len();
-        let capacity = args.capacity();
-        // the UDF dylib is responsible for freeing the memory of the args -- we leak it before
-        // calling the udf so that if it panics, we don't try to double-free the args
-        let ptr = args.leak();
-
-        let res = unsafe { self.udf.run(ptr.as_mut_ptr(), len, capacity).await };
-        let result_array = make_array(unsafe { from_ffi(res.0, &res.1).unwrap() });
-
-        ScalarValue::try_from_array(&result_array, 0)
-            .expect("could not get scalar from async UDF")
-    }
-}
 
 #[repr(C)]
 #[derive(Debug)]
@@ -344,61 +224,4 @@ pub struct FfiArraySchemaPair(FFI_ArrowArray, FFI_ArrowSchema);
 
 #[repr(C)]
 pub struct FfiArrayResult(FFI_ArrowArray, FFI_ArrowSchema, bool);
-
-impl ScalarUDFImpl for SyncUdfDylib {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> DFResult<DataType> {
-        Ok((*self.return_type).clone())
-    }
-
-    fn invoke(&self, args: &[ColumnarValue]) -> DFResult<ColumnarValue> {
-        let num_rows = args
-            .iter()
-            .map(|arg| {
-                if let ColumnarValue::Array(array) = arg {
-                    array.len()
-                } else {
-                    1
-                }
-            })
-            .max()
-            .unwrap();
- 
-        let mut args = args
-            .iter()
-            .map(|arg| {
-                let (array, schema) = to_ffi(&arg.clone().into_array(num_rows).unwrap().to_data()).unwrap();
-                FfiArraySchemaPair(array, schema)
-            })
-            .collect::<Vec<_>>();
-
-        let len = args.len();
-        let capacity = args.capacity();
-        // the UDF dylib is responsible for freeing the memory of the args -- we leak it before
-        // calling the udf so that if it panics, we don't try to double-free the args
-        let ptr = args.leak();
-
-        let FfiArrayResult(result_array, result_schema, valid) =
-            unsafe { (self.udf.run)(ptr.as_mut_ptr(), len, capacity) };
-
-        if !valid {
-            panic!("panic in UDF {}", self.name);
-        }
-
-        let result_array = unsafe { from_ffi(result_array, &result_schema).unwrap() };
-
-        Ok(ColumnarValue::Array(make_array(result_array)))
-    }
-}
