+++ crates/arroyo-udf/arroyo-udf-macros/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "arroyo-udf-macros"
+version = "0.10.0-dev"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+arrow-schema = "50.0.0"
+arroyo-udf-common = { path = "../arroyo-udf-common" }
+syn = {version = "2", features = ["full"]}
+quote = "1.0"
+proc-macro2 = "1.0"
