+++ crates/arroyo-udf/arroyo-udf-plugin/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod async_udf;
+
+pub use arroyo_udf_common::{FfiArraySchema, FfiArrays, RunResult, ArrowDatum};
